% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COBRE.data}
\alias{COBRE.data}
\title{fMRI brain networks of the COBRE dataset}
\format{\code{COBRE.data} is a list with two elements
\describe{
  \item{X.cobre}{A matrix containing the upper triangular part of the networks. Each row represents a subject,
  and the columns represent edges. The edge weights represent the Fisher-transformed correlation 
  between the
  fMRI time series of the nodes. The networks are composed of 263 nodes and 34453 different edges.
  For a description of the preprocessing steps to obtain the network edge weights, see \insertCite{relion2017network;textual}{graphclass}.
  }.
  \item{Y.cobre}{Class labels of the subjects in the dataset. \code{Y=1} represents schizophrenia status.}
}}
\source{
\url{http://fcon_1000.projects.nitrc.org/indi/retro/cobre.html}
}
\usage{
COBRE.data
}
\description{
This dataset contains fMRI brain networks of 124  subjects
and corresponding class labels indicating schizophrenia status (70 healthy controls and 54
schizophrenic subjects).
}
\examples{
data(COBRE.data)

}
\references{
\insertRef{Aine2017}{graphclass}
}
\keyword{datasets}
