% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_predict.R
\name{construct_D}
\alias{construct_D}
\title{Constructor for D penalty matrix, use it for efficiency when running the classifier multiple times.}
\usage{
construct_D(nodes = 264)
}
\arguments{
\item{nodes}{Number of nodes in the network, by default is 264 (Power parcellation).}
}
\value{
A sparse D matrix
}
\description{
Constructor for D penalty matrix, use it for efficiency when running the classifier multiple times.
}
\examples{
D = construct_D(100)
}
