% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_predict.R
\name{gc_predict}
\alias{gc_predict}
\title{Predict function for graph classifier}
\usage{
gc_predict(gc, X, type = "class", Ytest)
}
\arguments{
\item{gc}{A trained graph classifier object}

\item{X}{A matrix containing rows with vectorized upper triangular adjacency matrices (column-major order)}

\item{type}{Indicates the type of response. class: predicted classes. prob: predicted probabilities. error: misclassification error}

\item{Ytest}{If type = "error", true classes to compare.}
}
\value{
A vector containing the predicted classes.
}
\description{
Predict function for graph classifier
}
\examples{
X = matrix(rnorm(100*34453), nrow = 100)
Y = 2*(runif(100) > 0.5) - 1
gc = graphclass(X, Y = Y)
Xtest = matrix(rnorm(100*34453), nrow = 100)
}
