% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{get_matrix}
\alias{get_matrix}
\title{Returns a matrix from a vectorized network}
\usage{
get_matrix(beta, type = "intersection")
}
\arguments{
\item{beta}{Vectorized adjacency matrix.}

\item{type}{Either intersection for undirected networks, union for directed.}
}
\value{
Adjacency matrix for a vectorized network
}
\description{
Returns a matrix from a vectorized network
}
