% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{graphclass-package}
\alias{graphclass}
\alias{graphclass-package}
\title{Generate R documentation from inline comments.}
\description{
Roxygen2 allows you to write documentation in comment blocks co-located
with code.
}
\details{
The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
\author{
\strong{Maintainer}: Jesus Daniel Arroyo Relion \email{jarroyor@umich.edu}

}
