% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphclass.R
\name{graphclass}
\alias{graphclass}
\title{Train a graph classifier using regularized logistic regression.}
\usage{
graphclass(X = NULL, Adj_list = NULL, Y = NULL, Xtest = NULL,
  Ytest = NULL, type = "intersection", lambda1 = NULL, lambda2 = NULL,
  lambda = 0, rho = 0, gamma = 1e-05, params = NULL, id = "",
  verbose = F, D = NULL)
}
\arguments{
\item{X}{A matrix with the training sample, in wich each row represents a vectorized (by column order) upper triangular part of a network.}

\item{Adj_list}{A list of of symmetric matrices with 0 diagonal for training the classifier}

\item{Y}{A vector containing the class labels of the training sample (for now only 2 classes are supported).}

\item{Xtest}{A optional test matrix.}

\item{Ytest}{Labels of test set.}

\item{type}{should be either "intersection", "union" or "fusion", only "intersection" is currently supported.}

\item{lambda}{penalty parameter \eqn{lambda}, by default is set to 0.}

\item{rho}{penalty parameter \eqn{rho} controlling sparsity, by default is set to 0.}

\item{gamma}{ridge parameter (for numerical purposes).}

\item{params}{A list containing threshold parameters for the algorithm (see details)}

\item{verbose}{whether output is printed}

\item{D}{matrix \eqn{D} of the penalty; precomputing it can save time.}
}
\value{
An object containing the trained graph classifier.
}
\description{
Train a graph classifier using regularized logistic regression.
}
\examples{
X = matrix(rnorm(100*34453), nrow = 100)
Y = 2*(runif(100) > 0.5) - 1
gc = graphclass(X, Y = Y)
gc$train_error
}
