% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\encoding{UTF-8}
\name{matrix_to_vec}
\alias{matrix_to_vec}
\title{Returns a vectorized adjacency matrix}
\usage{
matrix_to_vec(A, type = c("undirected", "directed"))
}
\arguments{
\item{A}{Adjacency matrix of a network.}

\item{type}{Parameter to specify whether the adjacency matrix represents an \code{undirected}
or \code{directed} network. Default is \code{undirected}.}
}
\value{
Vector represeingt an adjacency matrix
}
\description{
Returns a vectorized adjacency matrix
}
