% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{node_sparsity}
\alias{node_sparsity}
\title{Returns node sparsity of a vector}
\usage{
node_sparsity(beta)
}
\arguments{
\item{beta}{Vectorized adjacency matrix.}
}
\value{
Percentage of inactive nodes in the graph solution
}
\description{
Returns node sparsity of a vector
}
