% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.graphclass.r
\name{plot.graphclass}
\alias{plot.graphclass}
\title{Plot function for graph classifier.}
\usage{
\method{plot}{graphclass}(object, ...)
}
\arguments{
\item{object}{trained graphclass object}
}
\description{
Plots the adjacency matrix of the coefficients network
}
\examples{
X = matrix(rnorm(100*34453), nrow = 100)
Y = 2*(runif(100) > 0.5) - 1
gc = graphclass(X, Y = factor(Y))
plot(gc)
}
