% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_adjmatrix}
\alias{plot_adjmatrix}
\title{Plot a vectorized adjacency matrix.}
\usage{
plot_adjmatrix(beta, type = "intersection")
}
\arguments{
\item{beta}{Vectorized adjacency matrix. For undirected networks use only upper triangle in column-major order, for directed use both}

\item{type}{Either intersection for undirected networks, union for directed.}
}
\description{
Plot a vectorized adjacency matrix.
}
\examples{
B = runif(34453)
plot_adjmatrix(B)
}
