% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_square_adj_mat}
\alias{plot_square_adj_mat}
\title{Plot a vectorized adjacency matrix with cells divisions}
\usage{
plot_square_adj_mat(edge_values, communities = NULL, type = "real",
  community_labels = c(1:13, -1), main = "", cut_at, sel_cells)
}
\arguments{
\item{edge_values}{Vectorized adjacency matrix. Only undirected networks are supported for now.}

\item{communities}{Community of each node}

\item{type}{Either "real" for valued networks, "prob" for [0,1] valued networks or "prob_cells" for equal value on each cell}

\item{community_labels}{Name of each community that will appear on the plot.}

\item{main}{Title of the plot}
}
\description{
Plot a vectorized adjacency matrix with cells divisions
}
