% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power.parcellation}
\alias{power.parcellation}
\title{Node assignments of the Power parcellation}
\format{
\code{power.parcellation} is a data frame in which the nodes represent nodes. The data frame has three columns:
\describe{
  \item{Master.Assignment}{Brain system number}
  \item{Color}{Color representing the system.}
  \item{Suggested.System}{Suggested brain system.}
  
}
}
\source{
\url{http://www.nil.wustl.edu/labs/petersen/Resources_files/Consensus264.xls}
}
\usage{
power.parcellation
}
\description{
Table containing the node assignments to brain systems of the 264 regions of interest from \insertCite{power2011functional;textual}{graphclass}. 
These nodes were used to construct COBRE.data and UMich.data. Note that node 75 is missing in the COBRE data.
}
\examples{
data(power.parcellation)

}
\references{
\insertRef{power2011functional}{graphclass}
}
\keyword{datasets}
