% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.graphclass.r
\name{predict.graphclass}
\alias{predict.graphclass}
\title{Predict function for graph classifier.}
\usage{
\method{predict}{graphclass}(object, newdata, type = "class", Ytest, ...)
}
\arguments{
\item{object}{trained graphclass object}

\item{newdata}{matrix of observations to predict. Each row corresponds to a new observation.}

\item{type}{type of response. class: predicted classes. prob: predicted probabilities. error: misclassification error}

\item{Ytest}{if type = "error", true classes to compare.}
}
\value{
A vector containing the predicted classes.
}
\description{
Predict function for graph classifier.
}
\examples{
X = matrix(rnorm(100*34453), nrow = 100)
Y = 2*(runif(100) > 0.5) - 1
gc = graphclass(X, Y = factor(Y))
Xtest = matrix(rnorm(100*34453), nrow = 100)
predictions = predict(gc, Xtest)
}
