% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COBRE.data}
\alias{COBRE.data}
\title{fMRI brain networks of the COBRE dataset}
\format{\code{COBRE.data} is a list with two elements
\describe{
  \item{X.cobre}{A matrix containing the upper triangular part of networks. Each row represents a subject,
  and the columns represent edges. The edge weights are calculated based on the correlation networks of the
  fMRI time series, after which a rank transformation among all weights in the network and a normalization
  between subjects is computed. The networks are composed of 263 nodes and 34453 different edges.
  For a description of the preprocessing steps to obtain the network edge weights, see \insertCite{relion2017network;textual}{graphclass}.
  }
  \item{Y.cobre}{Class labels of the subjects in the dataset. \code{Y=1} represents schizophrenia status.}
}}
\source{
\url{http://fcon_1000.projects.nitrc.org/indi/retro/cobre.html}
}
\usage{
COBRE.data
}
\description{
The COBRE dataset contains the fMRI brain networks of 124  subjects
and class labels of schizophrenia brain disease status.
}
\keyword{datasets}
