% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\encoding{UTF-8}
\name{get_matrix}
\alias{get_matrix}
\title{Returns a matrix from a vectorized network}
\usage{
get_matrix(beta, type = c("undirected", "directed"))
}
\arguments{
\item{beta}{Vectorized adjacency matrix.}

\item{type}{Parameter to specify whether the vector represents an \code{undirected}
or \code{directed} network. Default is \code{undirected}.}
}
\value{
Adjacency matrix for a vectorized network

# Obtain the adjacency matrix of a COBRE data subject
data(COBRE.data)
A <- get_matrix(COBRE.data$X.cobre[1,])
}
\description{
Returns a matrix from a vectorized network
}
