% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\encoding{UTF-8}
\name{matrix_to_vec}
\alias{matrix_to_vec}
\title{Convert matrix to vector.}
\usage{
matrix_to_vec(A, type = c("undirected", "directed"))
}
\arguments{
\item{A}{Adjacency matrix of a network.}

\item{type}{Parameter to specify whether the adjacency matrix represents an \code{undirected}
or \code{directed} network. Default is \code{undirected}.}
}
\value{
A vector containing the upper triangular part of an
 adjacency matrix (if the graph is undirected) or adjacency entry
 values ordered by column and excluding the diagonal entries (if the graph
 is directed).
}
\description{
The function encodes an adjacency matrix into a vector.
}
