% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{node_sparsity}
\alias{node_sparsity}
\title{Calculates the node sparsity of a vectorized adjacency matrix}
\usage{
node_sparsity(beta)
}
\arguments{
\item{beta}{Vectorized adjacency matrix.}
}
\value{
Percentage of inactive nodes in the graph solution
}
\description{
Calculates the node sparsity of a vectorized adjacency matrix
}
