% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\encoding{UTF-8}
\name{plot_adjmatrix}
\alias{plot_adjmatrix}
\title{Plots a vectorized adjacency matrix.}
\usage{
plot_adjmatrix(
  edgevalues,
  type = c("undirected", "directed"),
  edgetype = c("real", "prob", "binary"),
  communities = NULL,
  community_labels = NULL,
  main = "",
  axislabel = "Nodes",
  colorlims = NULL
)
}
\arguments{
\item{edgevalues}{Edge values of the adjacency matrix. The argument can be either a square adjacency matrix, or a 
vectorized adjacency matrix. For undirected networks, the vector should
contain the upper triangular part in column-major order. For directed networks use both}

\item{type}{If \code{edgevalues} is a vector, this parameter specifies whether the vector represents 
an \code{undirected} or \code{directed} network. Default is \code{undirected}.}

\item{edgetype}{Specify the type of edge values. For real-valued edges, \code{type = "real"}.
If the edges are between 0 and 1, \code{tipe = "prob"}. For binary edges, \code{type = "binary"}.}

\item{communities}{Optional argument to specify a list in which each element contains an array indicating
the indexes of the nodes on each community.}

\item{community_labels}{Labels for each community. The array should have the same length than \code{communities}.}

\item{main}{Title of the plot}

\item{axislabel}{Label for the axes.}

\item{colorlims}{An array with two elements indicating the minimum and maximum value in the color bar}
}
\value{
An objected returned by the function \code{levelplot} of the \code{lattice} package.
}
\description{
Draws a plot of the adjacency matrix represented by a vector
using the function \code{levelplot} of the \code{lattice} package.
}
\examples{

# Plot the adjacency matrix of a COBRE data subject
data(COBRE.data)
X1 <- COBRE.data$X.cobre[1,]

# Plot adjacency matrix with nodes in default order
plot_adjmatrix(X1)

# Plot adj. matrix divided by communities
data(power.parcellation)
# Node assignments (note that node 75 is missing on COBRE)
node.assignments <- power.parcellation$Master.Assignment[-75]
communities = lapply(c(1:13, -1), function(x) which(node.assignments==x))
plot_adjmatrix(X1, communities = communities, 
      community_labels = c(1:13, -1), axislabel = "Brain systems")

}
