% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.graphclass.r
\name{predict.graphclass}
\alias{predict.graphclass}
\title{Prediction function for graph classifier}
\usage{
\method{predict}{graphclass}(object, newdata, type = c("class", "prob", "error"), Ytest, ...)
}
\arguments{
\item{type}{Indicates the type of response: \code{class} for class predictions, \code{prob} for predicted probabilities, 
\code{error} for misclassification error (if \code{Ytest} is provided).}

\item{Ytest}{If type = "error", true classes to compare.}

\item{gc}{A trained graph classifier object}

\item{X}{A matrix containing rows with vectorized upper triangular adjacency matrices (column-major order)}
}
\value{
A vector containing the predicted classes.
}
\description{
Given an object of type \code{graphclass}, this function obtains the fitted classes for a new data.
}
\examples{
data(COBRE.data)
X <- COBRE.data$X.cobre
Y <- COBRE.data$Y.cobre

#Split into train and test
test <- runif(length(Y)) <= 0.1
gc <- graphclass(X = X[!test, ], Y = factor(Y[!test]), type = "intersection",
               lambda = 1e-4, rho = 1, gamma = 1e-5)
               
gc.test <- predict(gc, X[test, ]) 
predict(gc, X[test, ], "error", Ytest = Y[test])
}
